Updated tests and removed blankquestionnaireIT
