+++ src/test/java/uk/gov/ons/census/casesvc/messaging/ReceiptReceiverIT.java
@@ -29,7 +29,11 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.transaction.annotation.Transactional;
 import uk.gov.ons.census.casesvc.model.dto.CollectionCase;
+import uk.gov.ons.census.casesvc.model.dto.EventDTO;
 import uk.gov.ons.census.casesvc.model.dto.EventTypeDTO;
+import uk.gov.ons.census.casesvc.model.dto.FieldWorkFollowup;
+import uk.gov.ons.census.casesvc.model.dto.PayloadDTO;
+import uk.gov.ons.census.casesvc.model.dto.ResponseDTO;
 import uk.gov.ons.census.casesvc.model.dto.ResponseManagementEvent;
 import uk.gov.ons.census.casesvc.model.dto.UacDTO;
 import uk.gov.ons.census.casesvc.model.entity.Case;
@@ -50,6 +54,7 @@ public class ReceiptReceiverIT {
   private static final EasyRandom easyRandom = new EasyRandom();
   private final String TEST_NON_CCS_QID_ID = "1234567890123456";
   private final String TEST_CCS_QID_ID = "7134567890123456";
+  private final String TEST_QID_ID = "1234567890123456";
   private static final String TEST_UAC = easyRandom.nextObject(String.class);
 
   @Value("${queueconfig.receipt-response-inbound-queue}")
@@ -64,6 +69,9 @@ public class ReceiptReceiverIT {
   @Value("${queueconfig.rh-uac-queue}")
   private String rhUacQueue;
 
+  @Value("${queueconfig.fieldwork-followup-queue}")
+  private String fieldWorkFollowupQueue;
+
   @Autowired private RabbitQueueHelper rabbitQueueHelper;
   @Autowired private CaseRepository caseRepository;
   @Autowired private EventRepository eventRepository;
@@ -76,6 +84,7 @@ public class ReceiptReceiverIT {
     rabbitQueueHelper.purgeQueue(actionQueue);
     rabbitQueueHelper.purgeQueue(rhCaseQueue);
     rabbitQueueHelper.purgeQueue(rhUacQueue);
+    rabbitQueueHelper.purgeQueue(fieldWorkFollowupQueue);
     eventRepository.deleteAllInBatch();
     uacQidLinkRepository.deleteAllInBatch();
     caseRepository.deleteAllInBatch();
@@ -214,6 +223,146 @@ public class ReceiptReceiverIT {
     String utcDateAsString = new JSONObject(event.getEventPayload()).getString("dateTime");
     assertThat(isStringFormattedAsUTCDate(utcDateAsString)).isTrue();
   }
+
+  @Test
+  public void testPQRSSendReciptForQidThenQMSendBlankEvent()
+          throws InterruptedException, IOException {
+    // GIVEN
+    BlockingQueue<String> rhUacOutboundQueue = rabbitQueueHelper.listen(rhUacQueue);
+    BlockingQueue<String> rhCaseOutboundQueue = rabbitQueueHelper.listen(rhCaseQueue);
+    BlockingQueue<String> fieldworkFollowupQueue = rabbitQueueHelper.listen(fieldWorkFollowupQueue);
+
+    Case caze = setUpTestCase();
+    setUpUacQidLink(caze);
+
+    ResponseManagementEvent managementEvent = createValidReceiptEvent();
+
+    // WHEN
+    rabbitQueueHelper.sendMessage(inboundQueue, managementEvent);
+
+    // THEN
+    // check messages sent out after 1st unreceipted msg
+    boolean expectedReceipted = true;
+    checkMsgSentToRhCaseAndUac(
+            expectedReceipted, rhCaseOutboundQueue, rhUacOutboundQueue);
+    checkDatabaseUpdatedCorrectly(expectedReceipted, 1);
+
+    // Now send out a unreceipted event
+    ResponseManagementEvent unreceiptedEvent = createBlankQuestionaireEvent();
+    rabbitQueueHelper.sendMessage(inboundQueue, unreceiptedEvent);
+
+    checkMsgSentToRhCaseAndUac(
+            false, rhCaseOutboundQueue, rhUacOutboundQueue);
+
+    // Finally a new Event we're really interested in, a fieldworkFollowup should be sent out to
+    // fieldworkAdapter
+    FieldWorkFollowup fieldWorkFollowup =
+            rabbitQueueHelper.checkExpectedMessageReceived(
+                    fieldworkFollowupQueue, FieldWorkFollowup.class);
+
+    assertThat(fieldWorkFollowup.getCaseId()).isEqualTo(TEST_CASE_ID.toString());
+    assertThat(fieldWorkFollowup.getBlankQreReturned()).isTrue();
+
+    // Also check all the database gubbins again
+    expectedReceipted = false;
+    checkDatabaseUpdatedCorrectly(expectedReceipted, 2);
+  }
+
+  private void checkDatabaseUpdatedCorrectly(boolean expectedReceipted, int expectedEventCount) {
+    Case actualCase = caseRepository.findByCaseId(TEST_CASE_ID).get();
+    assertThat(actualCase.isReceiptReceived()).isEqualTo(expectedReceipted);
+
+    // check database for log eventDTO
+    List<Event> events = eventRepository.findAll();
+    assertThat(events.size()).isEqualTo(expectedEventCount);
+    Event event = events.get(expectedEventCount - 1);
+    assertThat(event.getEventDescription()).isEqualTo(QID_RECEIPTED);
+
+    UacQidLink actualUacQidLink = event.getUacQidLink();
+    assertThat(actualUacQidLink.getQid()).isEqualTo(TEST_QID_ID);
+    assertThat(actualUacQidLink.getUac()).isEqualTo(TEST_UAC);
+    assertThat(actualUacQidLink.getCaze().getCaseId()).isEqualTo(TEST_CASE_ID);
+    assertThat(actualUacQidLink.isActive()).isFalse();
+  }
+
+  private void setUpUacQidLink(Case caze) {
+    UacQidLink uacQidLink = new UacQidLink();
+    uacQidLink.setId(UUID.randomUUID());
+    uacQidLink.setCaze(caze);
+    uacQidLink.setCcsCase(false);
+    uacQidLink.setQid(TEST_QID_ID);
+    uacQidLink.setUac(TEST_UAC);
+    uacQidLinkRepository.saveAndFlush(uacQidLink);
+  }
+
+  private Case setUpTestCase() {
+    Case caze = easyRandom.nextObject(Case.class);
+    caze.setCaseId(TEST_CASE_ID);
+    caze.setReceiptReceived(false);
+    caze.setCcsCase(false);
+    caze.setUacQidLinks(null);
+    caze.setEvents(null);
+    caze = caseRepository.saveAndFlush(caze);
+
+    return caze;
+  }
+
+  private ResponseManagementEvent createBlankQuestionaireEvent() {
+    return createResponseReceivedUnreceiptedEvent(true);
+  }
+
+  private ResponseManagementEvent createValidReceiptEvent() {
+    return createResponseReceivedUnreceiptedEvent(false);
+  }
+
+  private ResponseManagementEvent createResponseReceivedUnreceiptedEvent(boolean unreceiped) {
+    ResponseManagementEvent managementEvent = new ResponseManagementEvent();
+
+    EventDTO event = new EventDTO();
+    event.setType(EventTypeDTO.RESPONSE_RECEIVED);
+    event.setSource("RECEIPT_SERVICE");
+    event.setChannel("PQRS");
+    managementEvent.setEvent(event);
+
+    ResponseDTO responseDTO = new ResponseDTO();
+    responseDTO.setQuestionnaireId(TEST_QID_ID);
+    responseDTO.setUnreceipt(unreceiped);
+
+    PayloadDTO payloadDTO = new PayloadDTO();
+    payloadDTO.setResponse(responseDTO);
+    managementEvent.setPayload(payloadDTO);
+
+    managementEvent.getEvent().setTransactionId(UUID.randomUUID());
+
+    return managementEvent;
+  }
+
+  private void checkMsgSentToRhCaseAndUac(
+          boolean receipted,
+          BlockingQueue<String> rhCaseOutboundQueue,
+          BlockingQueue<String> rhUacOutboundQueue)
+          throws IOException, InterruptedException {
+
+    // Check That Rh receives a case with corrected receipted value
+    CollectionCase actualCollectionCase =
+            rabbitQueueHelper
+                    .checkExpectedMessageReceived(rhCaseOutboundQueue)
+                    .getPayload()
+                    .getCollectionCase();
+    assertThat(actualCollectionCase.getId()).isEqualTo(TEST_CASE_ID.toString());
+    assertThat(actualCollectionCase.getReceiptReceived()).isEqualTo(receipted);
+
+    // Check that rh receive an uac_updated event
+    ResponseManagementEvent responseManagementEvent =
+            rabbitQueueHelper.checkExpectedMessageReceived(rhUacOutboundQueue);
+    assertThat(responseManagementEvent.getEvent().getType()).isEqualTo(EventTypeDTO.UAC_UPDATED);
+    UacDTO actualUacDTOObject = responseManagementEvent.getPayload().getUac();
+    assertThat(actualUacDTOObject.getUac()).isEqualTo(TEST_UAC);
+    assertThat(actualUacDTOObject.getQuestionnaireId()).isEqualTo(TEST_QID_ID);
+    assertThat(actualUacDTOObject.getCaseId()).isEqualTo(TEST_CASE_ID.toString());
+  }
+
+
 
   private boolean isStringFormattedAsUTCDate(String dateAsString) {
     try {
