+++ src/test/resources/application-test.yml
@@ -10,8 +10,6 @@ uacservice:
     port: 18164
 
 queueconfig:
-  fieldwork-uacupdated-queue: FieldworkAdapter.uacUpdated
-  uacupdated-routingkey: event.uac.update
   fieldwork-followup-queue: Action.Field
 
 caserefgeneratorkey: rPr3qsjlU42hsDX5npfuCzlyF4UYfDTO
