+++ src/test/resources/definitions.json
@@ -195,16 +195,6 @@
         "x-dead-letter-exchange": "",
         "x-message-ttl": 2000
       }
-    },
-    {
-      "name": "FieldworkAdapter.uacUpdated",
-      "vhost": "/",
-      "durable": true,
-      "auto_delete": false,
-      "arguments": {
-        "x-dead-letter-exchange": "delayedRedeliveryExchange",
-        "x-dead-letter-routing-key": "FieldworkAdapter.uacUpdated"
-      }
     }
   ],
   "exchanges": [
@@ -252,14 +242,6 @@
       "destination": "case.action",
       "destination_type": "queue",
       "routing_key": "event.case.*",
-      "arguments": {}
-    },
-    {
-      "source": "events",
-      "vhost": "/",
-      "destination": "FieldworkAdapter.uacUpdated",
-      "destination_type": "queue",
-      "routing_key": "event.uac.update",
       "arguments": {}
     },
     {
